Bug 75108 NC Service Template must trim client and management security groups
This pull request fixes the Bug 75108. There are some issues where the
input parameter contains heading or trailling spaces. hence the scipts
are updated to trim the heading and trailing spaces in input parameters,
only the text input parameters are modified, passoword, Domain name etc
are consumed as it is.