fix a race during the network set up
The Kubelet started to get the node pod CIDR ends up trying to spawn a container
if there is one configured and waiting. This results in multiple vEthernet adapters
and the ifAlias being a list. Use the InterfaceIndex returned by Get-NetIPAddress
to workaround this as the new vEthernet adapter does not have an address and isn't
returned.